feat:deletion service for category,s-category imporoved with proper conditioning

+++ backend/src/services/category.service.js
@@ -1,4 +1,5 @@
 import Category from "../models/Category.js";
+import SubCategory from "../models/SubCategory.js";
 
 export const createCategoryService = async({name})=>{
     const existingCategory = await Category.findOne({name:{$regex:new RegExp(`${name.trim()}$`,"i")}})
@@ -37,6 +38,14 @@ export const deleteCategoryService = async(id)=>{
     const category = await Category.findById(id)
     if(!category){
         throw new Error("Category not found")
+    }
+
+    const hasSubCategories = await SubCategory.exists({
+        category:id,
+    })
+
+    if(hasSubCategories){
+        throw new Error("Category cannot be deleted as it has sub-categories")
     }
 
     await category.deleteOne()
